requirements.txt
advent-of-code-data==1.3.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
Pebble==5.0.5
pluggy==1.3.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
termcolor==2.4.0
tomli==2.0.1
tzlocal==5.2
urllib3==2.1.0
